# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Vitas Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/professions": {
            "get": {
                "tags": [
                    "Professions"
                ],
                "summary": "Get all professions",
                "description": "Get all professions",
                "operationId": "b2db98e233abd571ee7789886e5a396c",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Professions"
                ],
                "summary": "Create a new professions",
                "description": "Create a new professions",
                "operationId": "6c9391424070b3a97da10d2f29c63d60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/professions/{id}": {
            "put": {
                "tags": [
                    "Professions"
                ],
                "summary": "Update a new profession",
                "description": "Update a new profession",
                "operationId": "7604d9272db7db13a0401e0692650492",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the profession",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Professions"
                ],
                "summary": "Delete profession by id",
                "description": "Delete profession by id",
                "operationId": "bbe06875e6347f0e69d5e06d3aa787b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the profession",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/questions": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get all Questions",
                "description": "Get all Questions",
                "operationId": "afd730ad376ffcf70c719b40485aeb5d",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Questions"
                ],
                "summary": "Create a new question",
                "description": "Create a new question",
                "operationId": "12214a9637b493d9b0975ffbf3d3202b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fails!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/questions/{id}": {
            "get": {
                "tags": [
                    "Questions"
                ],
                "summary": "Get question by id",
                "description": "Get question by id",
                "operationId": "fef3558041b395e6474e2b7766586e84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Questions"
                ],
                "summary": "Delete question by id",
                "description": "Delete question by id",
                "operationId": "90b597403c9fd31dd0e3efd37038c5e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the question",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}